Updated api module to match the correct api calls
