+++ api/exercises/test/courseData.json
@@ -3,10 +3,11 @@
     "courseId": 1,
     "title": "Course title",
     "exerciseType": "transcription",
-    "language": "en",
+    "language": "es",
     "courseCategory": "",
     "question": "This is the question",
     "answer": "answer",
+    "audioPath": "/public/audio/es_exercise_1.mp3",
     "answerOptions": [
       "answer option 1",
       "answer option 2",
