+++ api/user/data.js
@@ -2,10 +2,7 @@ var fs = require("fs");
 var path = require("path");
 
 module.exports = {
-  getCourseData: function() {
-    var data = fs.readFileSync(__dirname + "/test/courseData.json");
-    var jsonData = JSON.parse(data);
-
-    return jsonData;
+  getTest: function() {
+    return "test";
   }
 }
