+++ api/user/index.js
@@ -6,7 +6,6 @@ var express = require('express');
 var path = require('path');
 var userRouter = express.Router();
 var User = require('./models/users');
-var Course = require('./models/courses');
 
 
 userRouter.get('/', function(req, res, next) {
@@ -92,6 +91,8 @@ userRouter.post('/registerPost', function(req, res, next){
       res.send(response);
     }else{
       //PENDING: log the user in and send them to dashboard
+      response.user_id = user._id;
+      response.username = user.username;
       response.err = 0;
       response.message = "Successfully created user"
 
@@ -101,39 +102,10 @@ userRouter.post('/registerPost', function(req, res, next){
 
 });
 
-
+//will gathers all of the information needed for the dashboard
 userRouter.get('/dashboard', function(req, res, next) {
     res.send("Dashboard Router");
 });
-//this will add some data to our course schema
-userRouter.get('/dashboard/coursesInit',function(req, res, next){
-  res.setHeader('Content-Type', 'application/json');
-  var courses = data.getCourseData(); //gets data from json
-
-  for(var i = 0; i < courses.length ; i++){
-    //tries to add it to the table
-    var response = {};
-    Course.create(courses[i],function(error, course){
-      if(error){
-        response.err = -1;
-        response.message = "Course already in db";
-        response.error = error;
-        console.log(response);
-      }else{
-        response.courseId = course.courseId;
-        response.title = course.title;
-        response.err = 0;
-        response.message = "Course added";
-        console.log(response);
-      }
-    });
-  }
-
-  response.message = "Courses Data is set";
-  res.send(response);
-});
-
-
 
 
 //export routes to /api/index.js
