+++ app/exercises/transcription/component.js
@@ -22,7 +22,7 @@ export class Transcription extends React.Component {
         super(props);
         this.state = {
             userInput: (this.state || {}).userInput || 'enter your answer',
-            user_answer: [] //array to keep user answer 
+            user_answer: [] //array to keep user answer
         };
         this.handleUserInputChange = this.handleUserInputChange.bind(this);
         this.handleSubmit = this.handleSubmit.bind(this);
@@ -32,7 +32,7 @@ export class Transcription extends React.Component {
         this.setState({userInput: event.target.value})
     }
 
-    handleSubmit(event) { 
+    handleSubmit(event) {
         //for reference https://stackoverflow.com/questions/42569899/storing-numbers-entered-via-an-input-form-into-an-array
         var answer = this.state.value;
         var allAnswer = this.state.user_answer.concat([userInput]); //stores user answer to be concatted to user_answer
@@ -42,7 +42,6 @@ export class Transcription extends React.Component {
     }
 
     render() {
-        console.log(this.state.user_answer);
         return <div className={classes('page_container', 'flex_container')}>
 
             <div className={classes('exercise_container', 'flex_container')}>
@@ -105,7 +104,7 @@ export class Transcription extends React.Component {
 
                             </ul>
 
-    
+
                         <ul className={classes('char_bar', 'es_chars', 'flex_container')}>
 
                             <li><a href="#" className={styles.char_button}>á</a></li>
@@ -126,7 +125,7 @@ export class Transcription extends React.Component {
                                     value='enter'
                                     className={styles.enter_button}
                                 />
-                                
+
                             </form>
 
                             <div className={styles.tooltip_feedback}>
